@@ -1,7 +1,7 @@
 package com.example.huayeloltool.model.game;
 
-import com.example.huayeloltool.enums.GameEnums;
 import com.example.huayeloltool.enums.CommonResp;
+import com.example.huayeloltool.enums.GameEnums;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
