@@ -3,72 +3,33 @@ package com.example.huayeloltool.monitor;
 import com.alibaba.fastjson2.JSON;
 import com.alibaba.fastjson2.JSONArray;
 import com.alibaba.fastjson2.JSONObject;
-import com.example.huayeloltool.enums.Constant;
 import com.example.huayeloltool.enums.GameEnums;
-import com.example.huayeloltool.model.*;
-import com.example.huayeloltool.model.base.BaseUrlClient;
+import com.example.huayeloltool.model.game.CustomGameSession;
+import com.example.huayeloltool.model.game.Matchmaking;
 import com.example.huayeloltool.service.GameSessionUpdateService;
 import com.example.huayeloltool.service.GameStateUpdateService;
-import com.example.huayeloltool.service.LcuApiService;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.*;
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
-import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.DisposableBean;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
 import java.util.Base64;
 import java.util.concurrent.TimeUnit;
 
-@Component
 @Slf4j
-public class GameFlowMonitor implements CommandLineRunner, DisposableBean {
+public class GameFlowMonitor {
 
-    private WebSocket webSocket;
+    public final OkHttpClient client;
+    private final GameStateUpdateService gameStateUpdateService;
+    private final GameSessionUpdateService gameSessionUpdateService;
 
-    @Resource
-    GameStateUpdateService gameStateUpdateService;
-    @Resource
-    LcuApiService lcuApiService;
-    @Resource
-    GameSessionUpdateService gameSessionUpdateService;
-    @Resource
-    @Qualifier(value = "unsafeOkHttpClient")
-    private OkHttpClient client;
-
-    @Override
-    public void run(String... args) {
-        Pair<Integer, String> lolClientApiInfo = lcuApiService.getLolClientApiInfo(Constant.LOL_UX_PROCESS_NAME);
-        if (lolClientApiInfo == null) {
-            log.error("LOL接口进程不存在！");
-            return;
-        }
-        try {
-            // 初始化url请求路径
-            BaseUrlClient instance = BaseUrlClient.getInstance();
-            instance.setPort(lolClientApiInfo.getLeft());
-            instance.setToken(lolClientApiInfo.getRight());
-
-            // 初始化当前召唤师信息
-            Summoner summoner = Summoner.setInstance(lcuApiService.getCurrSummoner());
-            if (summoner == null) {
-                log.error("获取当前召唤师信息失败！");
-                return;
-            }
-
-            // 初始化监听器
-            initGameFlowMonitor(instance.getPort(), instance.getToken());
-        } catch (Exception e) {
-            log.error("initGameFlowMonitor error", e);
-        }
+    // 构造函数注入
+    public GameFlowMonitor(OkHttpClient client, GameStateUpdateService gameStateUpdateService, GameSessionUpdateService gameSessionUpdateService) {
+        this.client = client;
+        this.gameStateUpdateService = gameStateUpdateService;
+        this.gameSessionUpdateService = gameSessionUpdateService;
     }
-
 
     public void initGameFlowMonitor(int port, String token) throws Exception {
         String auth = Base64.getEncoder().encodeToString(("riot:" + token).getBytes());
@@ -77,27 +38,27 @@ public class GameFlowMonitor implements CommandLineRunner, DisposableBean {
                 .addHeader("Authorization", "Basic " + auth)
                 .build();
 
-        webSocket = client.newWebSocket(request, new WebSocketListener() {
+        client.newWebSocket(request, new WebSocketListener() {
             @SneakyThrows
             @Override
-            public void onOpen(@NotNull WebSocket webSocket, @NotNull Response response) {
+            public void onOpen( WebSocket webSocket,  Response response) {
                 log.info("Connected to LCU");
                 webSocket.send("[5, \"OnJsonApiEvent\"]");
             }
 
             @Override
-            public void onMessage(@NotNull WebSocket webSocket, @NotNull String text) {
+            public void onMessage( WebSocket webSocket,  String text) {
                 handleWebSocketMessage(text);
             }
 
             @Override
-            public void onFailure(@NotNull WebSocket webSocket, @NotNull Throwable t, Response response) {
+            public void onFailure( WebSocket webSocket,  Throwable t, Response response) {
                 log.error("WebSocket error: ", t);
             }
         });
 
         // 保持连接
-        boolean result = client.dispatcher().executorService().awaitTermination(Long.MAX_VALUE, TimeUnit.NANOSECONDS);
+        boolean b = client.dispatcher().executorService().awaitTermination(Long.MAX_VALUE, TimeUnit.NANOSECONDS);
     }
 
     private void handleWebSocketMessage(String message) {
@@ -105,49 +66,39 @@ public class GameFlowMonitor implements CommandLineRunner, DisposableBean {
             if (StringUtils.isEmpty(message)) {
                 return;
             }
+
             JSONArray arr = JSON.parseArray(message);
-            if (arr.size() < 3 || !"OnJsonApiEvent".equals(arr.getString(1))) return;
+            if (arr.size() < 3 || !"OnJsonApiEvent".equals(arr.getString(1))) {
+                return;
+            }
 
             JSONObject event = arr.getJSONObject(2);
             String uri = event.getString("uri");
             Object data = event.get("data");
 
             switch (uri) {
-                case "/lol-gameflow/v1/gameflow-phase":
-                    new Thread(() -> gameStateUpdateService.onGameFlowUpdate(data.toString())).start();
-                    break;
-                case "/lol-champ-select/v1/session":
-                    new Thread(() -> gameSessionUpdateService.onChampSelectSessionUpdate(data.toString())).start();
-                    break;
-                case "/lol-lobby-team-builder/v1/matchmaking":
-                    // 更新游戏模式
-                    if (data == null) {
-                        return;
-                    }
-                    Matchmaking matchmaking = JSON.parseObject(data.toString(), Matchmaking.class);
-                    if (matchmaking != null && BooleanUtils.isTrue(matchmaking.getIsCurrentlyInQueue())) {
-                        Integer queueId = matchmaking.getQueueId();
-                        if (queueId != null && queueId > 0) {
-                            String modeName = GameEnums.GameQueueID.getGameNameMap(queueId);
-                            CustomGameSessionDetails.getInstance().setQueueId(queueId);
-                            log.info("当前游戏模式为：{}", modeName);
-                        }
-                    }
-                    break;
+                case "/lol-gameflow/v1/gameflow-phase" ->
+                        new Thread(() -> gameStateUpdateService.onGameFlowUpdate(data.toString())).start();
+                case "/lol-champ-select/v1/session" ->
+                        new Thread(() -> gameSessionUpdateService.onChampSelectSessionUpdate(data.toString())).start();
+                case "/lol-lobby-team-builder/v1/matchmaking" -> handGameMode(data);
             }
         } catch (Exception e) {
             log.error("handleWebSocketMessage error", e);
         }
     }
 
-
-    @Override
-    public void destroy() {
-        if (webSocket != null) {
-            webSocket.close(1000, "Application shutdown");
-        }
-        if (client != null) {
-            client.dispatcher().executorService().shutdown();
+    private void handGameMode(Object data) {
+        if (data != null) {
+            Matchmaking matchmaking = JSON.parseObject(data.toString(), Matchmaking.class);
+            if (BooleanUtils.isTrue(matchmaking.getIsCurrentlyInQueue())) {
+                Integer queueId = matchmaking.getQueueId();
+                if (queueId != null && queueId > 0) {
+                    String modeName = GameEnums.GameQueueID.getGameNameMap(queueId);
+                    CustomGameSession.getInstance().setQueueId(queueId);
+                    log.info("当前游戏模式为：{}", modeName);
+                }
+            }
         }
     }
 
