@@ -1,27 +1,26 @@
 package com.example.huayeloltool.service;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.TypeReference;
+
+import com.alibaba.fastjson2.JSON;
+import com.alibaba.fastjson2.TypeReference;
+import com.example.huayeloltool.config.CommonBean;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.ResponseBody;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.net.SocketTimeoutException;
 
 @Slf4j
-@Component
 public class CommonRequest {
 
     public final OkHttpClient client;
 
-    public CommonRequest(@Qualifier(value = "unsafeOkHttpClient") OkHttpClient client) {
-        this.client = client;
+    public CommonRequest() {
+        CommonBean commonBean = new CommonBean();
+        this.client = commonBean.unsafeOkHttpClient();
     }
 
     public <T> T sendRequest(Request request, Class<T> responseClass) throws IOException {
